feat: update `tried` & `Trie` to receive options instead of arguments
Change the method signature for `tried` and `Tried` to receive
single argument `options` instead of array of string arguments
to be invoked by `add`.

Set the default options and allow it to be overridden given the
first argument. Currently, the options are for the `endKey` and
`endValue` (see constants).

Although this is technically a breaking change, a major version
bump is not required since this package hasn't been released or
published.